requirements.txt
pdfplumber
beautifulsoup4
requests
selenium-wire
python-docx
pandas
numpy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
psycopg2-binary
blinker==1.6.2
urllib3<2.0
gspread # <--- RÉ-AJOUTÉ ICI
